core/sim/rtl_sim/src-c/framework/requirements.txt
pycparser==2.20
pycparserext==2021.1
pyelftools==0.29
Jinja2==3.1.4
